Put MonadHashCons in the Covenant tree
